backend/requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.6
dnspython==2.3.0
pykerberos==1.2.4
pymongo==4.3.3
python-dotenv==0.21.1
python-snappy==0.6.1
sqlparse==0.4.3
djangorestframework==3.14.0
factory_boy==3.2.1